+++ partclone/src/reiser4clone.c
@@ -114,10 +114,10 @@ extern void initial_image_hdr(char* device, image_head* image_hdr)
     free_blocks = reiser4_format_get_free(fs->format);
     memcpy(image_hdr->magic, IMAGE_MAGIC, IMAGE_MAGIC_SIZE);
     memcpy(image_hdr->fs, reiser4_MAGIC, FS_MAGIC_SIZE);
-    image_hdr->block_size = get_ms_blksize(SUPER(fs->master));
-    image_hdr->totalblock = reiser4_format_get_len(fs->format);
-    image_hdr->usedblocks = reiser4_format_get_len(fs->format) - free_blocks;
-    image_hdr->device_size = (get_ms_blksize(SUPER(fs->master))*reiser4_format_get_len(fs->format));
+    image_hdr->block_size = (int)get_ms_blksize(SUPER(fs->master));
+    image_hdr->totalblock = (unsigned long long)reiser4_format_get_len(fs->format);
+    image_hdr->usedblocks = (unsigned long long)(reiser4_format_get_len(fs->format) - free_blocks);
+    image_hdr->device_size = (unsigned long long)((get_ms_blksize(SUPER(fs->master))*reiser4_format_get_len(fs->format)));
     fs_close();
 }
