+++ partclone/src/reiserfsclone.c
@@ -95,9 +95,9 @@ extern void initial_image_hdr(char* device, image_head* image_hdr)
     memcpy(image_hdr->magic, IMAGE_MAGIC, IMAGE_MAGIC_SIZE);
     memcpy(image_hdr->fs, reiserfs_MAGIC, FS_MAGIC_SIZE);
     image_hdr->block_size = (int)fs->super->s_v1.sb_block_size;
-    image_hdr->totalblock = (int)fs->super->s_v1.sb_block_count;
-    image_hdr->usedblocks = fs->super->s_v1.sb_block_count - fs->super->s_v1.sb_free_blocks;
-    image_hdr->device_size = (int)(fs->super->s_v1.sb_block_size*fs->super->s_v1.sb_block_count);
+    image_hdr->totalblock = (unsigned long long)fs->super->s_v1.sb_block_count;
+    image_hdr->usedblocks = (unsigned long long)fs->super->s_v1.sb_block_count - fs->super->s_v1.sb_free_blocks;
+    image_hdr->device_size = (unsigned long long)(fs->super->s_v1.sb_block_size*fs->super->s_v1.sb_block_count);
     fs_close();
 }
